+++ lib/pages/startmenue_seite.dart
@@ -111,7 +111,7 @@ class StartmenueSeite extends StatelessWidget {
                 const Spacer(),
                 const Center(
                   child: Text(
-                    'Web App 0.9.2 @ GitHub:',
+                    'Web App 0.9.3 @ GitHub:',
                     style: TextStyle(fontSize: 13, color: AppFarben.subtilerText),
                   ),
                 ),